refactor(spend): dedupe ledger row mappers, share EVM hash check
- Extract SESSION_COLS, CONVERSION_COLS, SPEND_TX_COLS and row mappers to lib/db/spend-ledger-rows.ts for spend-sessions and spend-admin
- Export isLedgerCanonicalEvmTxHash from spend-ledger-explorer-url; use in payment confirm normalizeTxHash
- Trim explorer URL JSDoc; replace PRD section symbols in JSDoc with plain text
- Add Vitest coverage for isLedgerCanonicalEvmTxHash
